TUM-29 - add repositories implementation and DBContext configuration
