--- Backend/Tumultu/Infrastructure/DataProviders/Dapper/Common/DapperRepositoryBase.cs
@@ -1,11 +1,0 @@
-namespace Tumultu.Infrastructure.DataProviders.Dapper.Common;
-
-internal abstract class DapperRepositoryBase
-{
-    protected IDbConnectionFactory ConnectionFactory { get; }
-
-    protected DapperRepositoryBase(IDbConnectionFactory connectionFactory)
-    {
-        ConnectionFactory = connectionFactory;
-    }
-}
