--- Backend/Tumultu/Infrastructure/DataProviders/Dapper/IDbConnectionFactory.cs
@@ -1,8 +1,0 @@
-using System.Data;
-
-namespace Tumultu.Infrastructure.DataProviders.Dapper;
-
-internal interface IDbConnectionFactory
-{
-    IDbConnection CreateConnection();
-}
