--- Backend/Tumultu/Infrastructure/DataProviders/EFCore/Files/FileRepository.cs
@@ -1,23 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using Tumultu.Application.Files;
-using Tumultu.Domain.Entities;
-using Tumultu.Infrastructure.DataProviders.EFCore.Common;
-
-namespace Tumultu.Infrastructure.DataProviders.EFCore.Files;
-
-internal class FileRepository : EfCoreRepositoryBase<FileEntity, Guid>, IFileRepository
-{
-    public FileRepository(EfCoreDbContext context) : base(context)
-    {
-    }
-
-    public async Task<IEnumerable<FileEntity>> GetAllByAnySignature(string? md5Signature, string? sha1Signature, string? sha256Signature)
-    {
-        return await Context.Set<FileEntity>()
-            .Where(file =>
-                file.MD5Signature == md5Signature
-                || file.SHA1Signature == sha1Signature
-                || file.SHA256Signature == sha256Signature)
-            .ToListAsync();
-    }
-}
