--- Backend/Tumultu/Infrastructure/DataProviders/EFCore/Users/UserRepository.cs
@@ -1,12 +1,0 @@
-using Tumultu.Application.Users;
-using Tumultu.Domain.Entities;
-using Tumultu.Infrastructure.DataProviders.EFCore.Common;
-
-namespace Tumultu.Infrastructure.DataProviders.EFCore.Users;
-
-internal class UserRepository : EfCoreRepositoryBase<User, Guid>, IUserRepository
-{
-    public UserRepository(EfCoreDbContext context) : base(context)
-    {
-    }
-}
