+++ Backend/Tumultu/Infrastructure/Database/Dapper/IDbConnectionFactory.cs
@@ -1,0 +1,8 @@
+using System.Data;
+
+namespace Tumultu.Infrastructure.Database.Dapper;
+
+internal interface IDbConnectionFactory
+{
+    IDbConnection CreateConnection();
+}
