+++ Backend/Tumultu/Infrastructure/Database/EFCore/Configuration/FileEntityTypeConfiguration.cs
@@ -1,0 +1,15 @@
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Tumultu.Domain.Entities;
+
+namespace Tumultu.Infrastructure.Database.EFCore.Configuration;
+
+internal class FileEntityTypeConfiguration : IEntityTypeConfiguration<FileEntity>
+{
+    public void Configure(EntityTypeBuilder<FileEntity> builder)
+    {
+        builder.OwnsMany(entity => entity.FileStrings);
+        
+        //TODO add property constraints and indexes (ex. like max length for string values)
+    }
+}
