+++ Backend/Tumultu/Infrastructure/Infrastructure.csproj
@@ -10,12 +10,12 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.0"/>
-    <PackageReference Include="Dapper" Version="2.1.24" />
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.0" />
-  </ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="8.0.0"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.0"/>
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.0"/>
+    <PackageReference Include="Dapper" Version="2.1.24"/>
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0"/>
 
-  <ItemGroup>
     <ProjectReference Include="..\Application\Application.csproj"/>
   </ItemGroup>
